DiscordBot/requirements.txt
discord
app
socketio
datetime
json
os
threading
flask_socketio